input: fix ili7807s touch dislocation when aging test
Problem: aging test, touch dislocation.

Cause: during the boot process, when the rst pin is pulled, the int pin will also be pulled down, and the system respond
to the interrupt in advance, but at this time there is no firmware in the IC, which will trigger esd recovery to reload
the firmware. The resolution information is stored in the firmware, but the firmware path has not been assigned at this time,
and the firmware cannot be applied, resulting in the upgrade failure. If the resolution cannot be obtained, the default
resolution will be used, which will lead to incorrect resolution setting.

Solution: during the boot process, only after the firmware upgrade successfully, system can response to interrupts.

Change-Id: I0ad032de0b96ed86177a9aa4a16f9aa87613042d
Reviewed-on: https://gerrit.mot.com/2514709
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Jun Weng <wengjun1@motorola.com>
Reviewed-by: Yeqing Wang <wangyq13@motorola.com>
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key