wireless: Switch from EPP to BPP for CEC
For Wireless Charging California Energy Efficiency CEC Certification,
When DUT is charging FULL and enable_stop_epp is enable,
we need to switch from EPP to BPP.
And the CPS4038 will reboot when switch from EPP to BPP.
So we need to reset the online value in cps_wls_chrg_get_property
during the period of CPS4038 restart.
In this way,the charging animation will not be shown again.
And the DUT not reboot in charging only mode.
And clear stop_epp_flag when detected an attach event(wls stand out).
Generally, we need to set the mode select pin high.
There has 5000ms timeout if we putout DUT when switch to BPP.
If timeout, we need to clean stop_epp_flag and set mode_sel gpio high.

Change-Id: I25359f5a8e4909c27695d2dfe8a5bd383e606756
Signed-off-by: Guocheng Wang <wanggc6@motorola.com>
Reviewed-on: https://gerrit.mot.com/2558002
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Haijian Ma <mahj8@motorola.com>
Reviewed-by: Lu Chai <chailu1@motorola.com>
Reviewed-by: Jiacheng Ou <oujc1@motorola.com>
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key