wireless: Fix wls offset det err when thermal limit
Slow charging pops up and disappears soon when the thermal limit
wls charging.
Because when thermal limit,the wireless charging current is not
very accurate and stable.It's cannot be used as a basis to detection
of offset.
So remove iout check when EPP wls charging.

Change-Id: If530c30b84e2bd83128d128d7706d172aea36ac0
Signed-off-by: Guocheng Wang <wanggc6@motorola.com>
Reviewed-on: https://gerrit.mot.com/2633328
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Haijian Ma <mahj8@motorola.com>
Reviewed-by: Wei Xu <xuwei9@lenovo.com>
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key